Move helpers out of handlers
